Simplify adding basic title components.
